+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/view/CCv2DataView.kt
@@ -38,7 +38,7 @@ abstract class CCv2DataView<T : CCv2Dto> {
 
     fun fetchingInProgressPanel(subscriptions: Collection<CCv2Subscription>): DialogPanel = panel {
         subscriptions.forEach {
-            collapsibleGroup(it.toString()) {
+            collapsibleGroup(it.presentableName) {
                 row {
                     label("Fetching ${tab.title} data, Please wait...")
                         .align(Align.CENTER)