+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/view/CCv2DeploymentsDataView.kt
@@ -39,7 +39,7 @@ object CCv2DeploymentsDataView : CCv2DataView<CCv2DeploymentDto>() {
     override fun dataPanel(project: Project, data: Map<CCv2Subscription, Collection<CCv2DeploymentDto>>): DialogPanel = if (data.isEmpty()) noDataPanel()
     else panel {
         data.forEach { (subscription, builds) ->
-            collapsibleGroup(subscription.toString()) {
+            collapsibleGroup(subscription.presentableName) {
                 if (builds.isEmpty()) {
                     noData()
                 } else {