+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/view/CCv2EnvironmentDetailsView.kt
@@ -354,7 +354,7 @@ class CCv2EnvironmentDetailsView(
                 val environmentCode = if (environment.name != environment.code) "${environment.code} - ${environment.name}"
                 else environment.name
 
-                label("$subscription - $environmentCode")
+                label("${subscription.presentableName} - $environmentCode")
                     .comment("Environment")
                     .bold()
                     .component.also {