+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### `External Dependencies` enhancements
 - Re-enabled possibility to download new dependencies via IDE [#1593](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1593)
 
+### Fixed
+- Fixed CCv2 `subscription` and `s-user` representation [#1594](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1594)
+
 ## [2025.2.4.2]
 
 ### `xDebugger` enhancements